requirements.txt
Django==1.11.5
djangorestframework==3.7.7
Markdown==2.6.11
django-cors-headers
django-filter==1.1.0

django-simple-pagination==1.1.8
django-tastypie==0.14.0
djangorestframework-files==1.0.1
gunicorn==19.6.0
psycopg2==2.7.3.1
dj-database-url==0.4.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
six==1.11.0
whitenoise==3.1
django-rest-swagger
django-rest-auth
djangorestframework-jwt
django-allauth
boto
django-storages
pillow
pusher
